requirements.txt
django>=3.2.13,<3.3
requests>=2.25.0
pillow>=8.0
django-crispy-forms>=2.1.0
urllib3>=2.0.4
django-uni-form>=0.9.0
bootstrap==0.1.0
pip~=23.2.1
wheel~=0.41.1
Pillow>=8.0
setuptools~=68.0.0
asgiref~=3.7.2
sqlparse~=0.4.4
PyYAML~=6.0.1
Jinja2~=3.1.2
djangorestframework>=3.12.4,<3.13